Only run Web Sdk Interactive tests on net6.0 and net7.0
