+++ src/Dotnet.Script.Shared.Tests/InteractiveRunnerTestsBase.cs
@@ -17,7 +17,7 @@ namespace Dotnet.Script.Shared.Tests
             testOutputHelper.Capture();
         }
 
-        private class InteractiveTestContext
+        protected class InteractiveTestContext
         {
             public InteractiveTestContext(ScriptConsole console, InteractiveRunner runner)
             {
@@ -29,7 +29,7 @@ namespace Dotnet.Script.Shared.Tests
             public InteractiveRunner Runner { get; }
         }
 
-        private InteractiveTestContext GetRunner(params string[] commands)
+        protected InteractiveTestContext GetRunner(params string[] commands)
         {
             var reader = new StringReader(string.Join(Environment.NewLine, commands));
             var writer = new StringWriter();
@@ -306,26 +306,6 @@ namespace Dotnet.Script.Shared.Tests
             await ctx.Runner.RunLoop();
             var result = ctx.Console.Out.ToString();
             Assert.Contains("[Submission#0+SimpleTargets+TargetDictionary]", result);
-        }
-
-        [Fact]
-        public async Task ShouldCompileAndExecuteWithWebSdk()
-        {
-            var commands = new[]
-            {
-                @"#r ""sdk:Microsoft.NET.Sdk.Web""",
-                "using Microsoft.AspNetCore.Builder;",
-                "var a = WebApplication.Create();",
-                @"a.GetType()",
-                "#exit"
-            };
-
-            var ctx = GetRunner(commands);
-            await ctx.Runner.RunLoop();
-
-            var result = ctx.Console.Out.ToString();
-
-            Assert.Contains("[Microsoft.AspNetCore.Builder.WebApplication]", result);
         }
     }
 }
