+++ src/Dotnet.Script.Tests/InteractiveRunnerTests.cs
@@ -1,6 +1,7 @@
 ﻿using Xunit;
 using Dotnet.Script.Shared.Tests;
 using Xunit.Abstractions;
+using System.Threading.Tasks;
 
 namespace Dotnet.Script.Tests
 {
@@ -9,6 +10,26 @@ namespace Dotnet.Script.Tests
     {
         public InteractiveRunnerTests(ITestOutputHelper testOutputHelper) : base(testOutputHelper)
         {
+        }
+
+        [Fact]
+        public async Task ShouldCompileAndExecuteWithWebSdk()
+        {
+            var commands = new[]
+            {
+                @"#r ""sdk:Microsoft.NET.Sdk.Web""",
+                "using Microsoft.AspNetCore.Builderss;",
+                "var a = WebApplication.Create();",
+                @"a.GetType()",
+                "#exit"
+            };
+
+            var ctx = GetRunner(commands);
+            await ctx.Runner.RunLoop();
+
+            var result = ctx.Console.Out.ToString();
+
+            Assert.Contains("[Microsoft.AspNetCore.Builder.WebApplication]", result);
         }
     }
 }
